Remove unnecessary cmake includes and variables
oxygen-sounds doesn't compile anything and so it doesn't need to know
about or check for compiler settings.
It also doesn't need Qt or searches it, so no need to have a minimal
version for it.
It think the only useful thing from KDECMakeSettings in this case is
an uninstall target.